remove WrapInLambdaExpression
