+++ RecordParser/Builders/Writer/VariableLengthWriterBuilder.cs
@@ -2,6 +2,7 @@
 using RecordParser.Parsers;
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Globalization;
 using System.Linq;
 using System.Linq.Expressions;
@@ -64,7 +65,7 @@ namespace RecordParser.Builders.Writer
     public class VariableLengthWriterBuilder<T> : IVariableLengthWriterBuilder<T>
     {
         private readonly Dictionary<int, MappingWriteConfiguration> list = new();
-        private readonly Dictionary<Type, Func<Expression, Expression, Expression, Expression>> dic = new();
+        private readonly Dictionary<Type, Delegate> dic = new();
 
         /// <summary>
         /// Customize configuration for individual member.
@@ -83,10 +84,24 @@ namespace RecordParser.Builders.Writer
         public IVariableLengthWriterBuilder<T> Map<R>(Expression<Func<T, R>> ex, int indexColumn, FuncSpanTIntBool<R> converter = null)
         {
             var member = ex.Body;
-            var config = new MappingWriteConfiguration(member, indexColumn, null, converter.WrapInLambdaExpression(), null, default, default, typeof(R));
+            var config = new MappingWriteConfiguration(member, indexColumn, null, converter, null, default, default, typeof(R));
             list.Add(indexColumn, config);
             return this;
         }
+
+        //public IVariableLengthWriterBuilder<T> Map(Expression<Func<T, string>> ex, int indexColumn, FuncSpanTIntBool converter)
+        //{
+        //    var member = ex.Body;
+        //    var config = new MappingWriteConfiguration(member, indexColumn, null, converter == null ? null : new FuncSpanTIntBool<string>(
+        //        (span, int) =>
+        //        {
+
+
+
+        //        }).WrapInLambdaExpression(), null, default, default, typeof(string));
+        //    list.Add(indexColumn, config);
+        //    return this;
+        //}
 
         /// <summary>
         /// Customize configuration for individual member.
@@ -116,7 +131,7 @@ namespace RecordParser.Builders.Writer
         /// <returns>An object to configure the mapping.</returns>
         public IVariableLengthWriterBuilder<T> DefaultTypeConvert<R>(FuncSpanTIntBool<R> ex)
         {
-            dic.Add(typeof(R), ex?.WrapInLambdaExpression());
+            dic.Add(typeof(R), ex);
             return this;
         }
 
@@ -136,6 +151,104 @@ namespace RecordParser.Builders.Writer
             var expression = VariableLengthWriterEngine.GetFuncThatSetProperties<T>(maps, cultureInfo);
 
             return new VariableLengthWriter<T>(expression.Compile(), separator);
+        }
+
+        /// <summary>
+        /// Write a string into span as quoted field when
+        /// 1) the string contains one or more quote character
+        /// 2) the string contains the column delimiter
+        /// </summary> 
+        private static bool Quote(string text, Span<char> span, out int written)
+        {
+            if (text.Length > span.Length)
+            {
+                written = 0;
+                return false;
+            }
+
+            char quote = '"';
+            ReadOnlySpan<char> separator = " , ";
+
+            var value = text.AsSpan();
+            var length = value.Length;
+            int i = 0;
+
+            var quoteFounds = 0;
+            var containsSeparator = false;
+
+            for (; i < length; i++)
+            {
+                if (value[i] == quote)
+                {
+                    quoteFounds++;
+                    continue;
+                }
+
+                if (containsSeparator == false && value.Slice(i).StartsWith(separator))
+                {
+                    containsSeparator = true;
+                }
+            }
+
+            if (quoteFounds == 0)
+            {
+                if (containsSeparator)
+                {
+                    if (text.Length + 2 > span.Length)
+                    {
+                        written = 0;
+                        return false;
+                    }
+                    else
+                    {
+                        span[0] = quote;
+                        value.CopyTo(span.Slice(1));
+                        span[value.Length + 1] = quote;
+
+                        written = text.Length + 2;
+                        return true;
+                    }
+                }
+                else
+                {
+                    value.CopyTo(span);
+                    written = value.Length;
+                    return true;
+                }
+            }
+            else
+            {
+                var newLengh = text.Length + quoteFounds + 2;
+
+                if (newLengh > span.Length)
+                {
+                    written = 0;
+                    return false;
+                }
+                else
+                {
+                    var j = 0;
+
+                    span[++j] = quote;
+
+                    for (i = 0; i < length; i++, j++)
+                    {
+                        span[j] = value[i];
+
+                        if (value[i] == quote)
+                        {
+                            span[++j] = quote;
+                        }
+                    }
+
+                    span[j] = quote;
+
+                    Debug.Assert(j == newLengh);
+
+                    written = newLengh;
+                    return true;
+                }
+            }
         }
     }
 }
